requirements.txt
django>=2.2,<2.3
psycopg2-binary>=2.8,<2.9
python-decouple>=3.3,<3.4
django-filter>=2.2,<2.3
